# docs/swagger.yaml
basePath: /
definitions:
...
  database.CreateJobParams:
    properties:
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      delivery:
        example:
        - 10
        - 20
        items:
          type: integer
        type: array
      location:
        $ref: '#/definitions/util.LocationParams'
      pickup:
        example:
        - 5
        - 15
        items:
          type: integer
        type: array
      priority:
        example: 10
        type: integer
      service:
        example: 120
        type: integer
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
    required:
    - location
    type: object
  database.CreateJobTimeWindowParams:
    properties:
      tw_close:
        example: "2021-12-31 23:59:00"
        type: string
      tw_open:
        example: "2021-12-31 23:00:00"
        type: string
    required:
    - tw_close
    - tw_open
    type: object
...
  database.Job:
    properties:
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      delivery:
        example:
        - 10
        - 20
        items:
          type: integer
        type: array
      id:
        example: "1234567812345678"
        type: string
      location:
        $ref: '#/definitions/util.LocationParams'
      pickup:
        example:
        - 5
        - 15
        items:
          type: integer
        type: array
      priority:
        example: 10
        type: integer
      project_id:
        example: "1234567812345678"
        type: string
      service:
        example: 120
        type: integer
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
  database.JobTimeWindow:
    properties:
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      id:
        example: "1234567812345678"
        type: string
      tw_close:
        example: "2021-12-31 23:59:00"
        type: string
      tw_open:
        example: "2021-12-31 23:00:00"
        type: string
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
...
  util.LocationParams:
    properties:
      latitude:
        example: 2.0365
        type: number
      longitude:
        example: 48.6113
        type: number
    required:
    - latitude
    - longitude
    type: object
  util.MultiError:
    properties:
      errors:
        example:
        - Error message1
        - Error message2
        items:
          type: string
        type: array
    type: object
  util.NotFound:
    properties:
      error:
        example: Not Found
        type: string
    type: object
...
  util.Success:
    properties:
      success:
        example: "true"
        type: string
    type: object
host: localhost:9100
info:
  contact:
    email: info@georepublic.de
    name: Team Georepublic
  description: This is an API for scheduling VRP tasks. Source code can be found on
    https://github.com/Georepublic/pg_scheduleserv
  license:
    name: GNU Affero General Public License
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://swagger.io/terms/
  title: pg_scheduleserv API
  version: 0.1.0
paths:
...
  /jobs/{job_id}:
    delete:
      consumes:
      - application/json
      description: Delete a job with its job_id
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete a job
      tags:
      - Job
    get:
      consumes:
      - application/json
      description: Fetch a job with its job_id
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.NotFound'
      summary: Fetch a job
      tags:
      - Job
    patch:
      consumes:
      - application/json
      description: Update a job (partial update) with its job_id
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Update a job
      tags:
      - Job
  /jobs/{job_id}/time_windows:
    get:
      consumes:
      - application/json
      description: Get a list of job time windows for a job with job_id
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.JobTimeWindow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List job time windows for a job
      tags:
      - Job
    post:
      consumes:
      - application/json
      description: Create a new job time window with the input payload
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      - description: Create job time window
        in: body
        name: JobTimeWindow
        required: true
        schema:
          $ref: '#/definitions/database.CreateJobTimeWindowParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.JobTimeWindow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new job time window
      tags:
      - Job
  /jobs/{job_id}/time_windows/{tw_open}/{tw_close}:
    delete:
      consumes:
      - application/json
      description: Delete job time windows for a job with job_id
      parameters:
      - description: Job ID
        in: path
        name: job_id
        required: true
        type: integer
      - description: Job opening Time Window
        in: path
        name: tw_open
        required: true
        type: string
      - description: Job closing Time Window
        in: path
        name: tw_close
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete job time windows
      tags:
      - Job
...
  /projects/{project_id}/jobs:
    get:
      consumes:
      - application/json
      description: Get a list of jobs for a project with project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List jobs for a project
      tags:
      - Job
    post:
      consumes:
      - application/json
      description: Create a new job with the input payload
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Job object
        in: body
        name: Job
        required: true
        schema:
          $ref: '#/definitions/database.CreateJobParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Job'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new job
      tags:
      - Job
...
schemes:
- http
- https
swagger: "2.0"